project restructured like it should
