+++ app/src/main/java/org/hvdw/fythwonekey/activities/OneKeyBTPhone.java
@@ -1,4 +1,4 @@
-package org.hvdw.fythwonekey;
+package org.hvdw.fythwonekey.activities;
 
 import android.Manifest;
 import android.app.Activity;
@@ -11,6 +11,10 @@ import android.content.ComponentName;
 import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
 import android.widget.Toast;
+
+import org.hvdw.fythwonekey.MySettings;
+import org.hvdw.fythwonekey.R;
+import org.hvdw.fythwonekey.Utils;
 
 
 public class OneKeyBTPhone extends Activity {
