+++ app/src/main/java/org/hvdw/fythwonekey/activities/OneKeyEQ.java
@@ -1,4 +1,4 @@
-package org.hvdw.fythwonekey;
+package org.hvdw.fythwonekey.activities;
 
 import android.Manifest;
 import android.app.Activity;
@@ -14,6 +14,10 @@ import android.view.KeyEvent;
 import android.widget.Toast;
 
 import android.media.AudioManager;
+
+import org.hvdw.fythwonekey.MySettings;
+import org.hvdw.fythwonekey.R;
+import org.hvdw.fythwonekey.Utils;
 
 
 public class OneKeyEQ extends Activity {
