+++ app/src/main/java/org/hvdw/fythwonekey/receivers/BTPhoneReceiver.java
@@ -1,4 +1,4 @@
-package org.hvdw.fythwonekey;
+package org.hvdw.fythwonekey.receivers;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
@@ -11,6 +11,7 @@ import android.widget.Toast;
 
 import org.hvdw.fythwonekey.MySettings;
 import org.hvdw.fythwonekey.R;
+import org.hvdw.fythwonekey.Utils;
 
 
 public class BTPhoneReceiver extends BroadcastReceiver {
